+++ projects/storefrontlib/src/layout/launch-dialog/services/index.ts
@@ -1,5 +1,5 @@
-export * from './inline-render-strategy.service';
+export * from './inline-render.strategy';
 export * from './launch-dialog.service';
 export * from './launch-render.strategy';
-export * from './outlet-render-strategy.service';
-export * from './routing-render-strategy.service';
+export * from './outlet-render.strategy';
+export * from './routing-render.strategy';
