+++ projects/storefrontlib/src/layout/launch-dialog/services/inline-render.strategy.spec.ts
@@ -6,7 +6,7 @@ import {
 } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { LaunchConfig, LaunchInlineDialog, LAUNCH_CALLER } from '../config';
-import { InlineRenderStrategy } from './inline-render-strategy.service';
+import { InlineRenderStrategy } from './inline-render.strategy';
 
 const testTemplate = {} as ComponentFactory<any>;
