+++ projects/storefrontlib/src/layout/launch-dialog/services/outlet-render.strategy.spec.ts
@@ -3,7 +3,7 @@ import { TestBed } from '@angular/core/testing';
 import { OutletPosition, OutletService } from '../../../cms-structure/index';
 import { OutletRendererService } from '../../../cms-structure/outlet/outlet-renderer.service';
 import { LaunchConfig, LaunchOutletDialog, LAUNCH_CALLER } from '../config';
-import { OutletRenderStrategy } from './outlet-render-strategy.service';
+import { OutletRenderStrategy } from './outlet-render.strategy';
 
 @Component({
   template: '',
