+++ projects/storefrontlib/src/layout/launch-dialog/services/routing-render.strategy.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { RoutingService } from '@spartacus/core';
 import { LaunchConfig, LaunchRoute, LAUNCH_CALLER } from '../config/index';
-import { RoutingRenderStrategy } from './routing-render-strategy.service';
+import { RoutingRenderStrategy } from './routing-render.strategy';
 
 const mockLaunchConfig: LaunchConfig = {
   launch: {
